Live Review 이전 코드 내용
