+++ src/components/Main/MainCarousel/MainCarousel.scss
@@ -1,11 +1,11 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 @mixin arrowBtn {
   position: absolute;
   width: 50px;
   height: 50px;
   border: none;
-  background-color: #ddd8d0;
+  background-color: #c7c2bb;
   opacity: 60%;
 }
 
@@ -48,7 +48,7 @@
     width: 6%;
     height: 6%;
     border-radius: 10%;
-    background-color: #ddd8d0;
+    background-color: #c7c2bb;
     text-align: center;
     opacity: 60%;
   }
