+++ src/components/Main/MainReview/ReviewCarousel.js
@@ -1,6 +1,6 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import React, { useEffect, useRef, useState } from 'react';
 import { faArrowLeft, faArrowRight } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import './ReviewCarousel.scss';
 
 const ReviewCarousel = () => {
