+++ src/components/Main/Menu/Menu.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 @mixin title {
   font-size: 40px;
   font-weight: bold;
