+++ src/components/Main/Menu/MenuInfo.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 @mixin itemText {
   padding-top: 10px;
   font-size: 20px;
