--- src/components/Main/NewProducts.scss
@@ -1,54 +1,0 @@
-@import '../../styles/variables.scss';
-@mixin itemText {
-  padding-top: 10px;
-  font-size: 20px;
-  font-weight: bold;
-}
-
-.newProducts {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  gap: 100px 35px;
-  margin-top: 30px;
-
-  .newProduct {
-    .productImgBox {
-      .productImg {
-        width: 100%;
-        height: 350px;
-      }
-    }
-
-    .productContent {
-      @include itemText();
-      display: flex;
-      justify-content: space-between;
-      padding-top: 20px;
-
-      .productInfo {
-        .productName {
-          padding-bottom: 15px;
-        }
-
-        .productPrice {
-          .productServings {
-            font-size: 15px;
-          }
-        }
-      }
-
-      .productCart {
-        width: 50px;
-        height: 50px;
-        border: none;
-        border-radius: 50%;
-        background-color: $mainColor;
-        cursor: pointer;
-
-        .cartIcon {
-          font-size: 18px;
-        }
-      }
-    }
-  }
-}
