+++ src/components/Main/Recommend/Recommend.js
@@ -1,6 +1,6 @@
-import React, { useState, useEffect, useRef } from 'react';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faArrowLeft, faArrowRight } from '@fortawesome/free-solid-svg-icons';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import React, { useEffect, useRef, useState } from 'react';
 import RecommendDrop from './RecommendDrop.js';
 import './Recommend.scss';
