+++ src/components/Main/Recommend/Recommend.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .recommend {
   @include flexCenter;
