+++ src/components/Main/Recommend/RecommendDrop.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../styles/variables.scss';
 
 .recommendDrop {
   position: relative;
